+++ app/src/main/java/open/furaffinity/client/adapter/journalSectionsPagerAdapter.java
@@ -10,9 +10,9 @@ import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
 
 import open.furaffinity.client.R;
-import open.furaffinity.client.fragmentTabs.comments;
-import open.furaffinity.client.fragmentTabs.notImplementedYet;
-import open.furaffinity.client.fragmentTabs.webViewContent;
+import open.furaffinity.client.fragmentTabsNew.comments;
+import open.furaffinity.client.fragmentTabsOld.notImplementedYet;
+import open.furaffinity.client.fragmentTabsOld.webViewContent;
 import open.furaffinity.client.pages.journal;
 import open.furaffinity.client.utilities.messageIds;
