+++ app/src/main/java/open/furaffinity/client/adapter/msgPmsMessageSectionsPagerAdapter.java
@@ -10,8 +10,8 @@ import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
 
 import open.furaffinity.client.R;
-import open.furaffinity.client.fragmentTabs.notImplementedYet;
-import open.furaffinity.client.fragmentTabs.webViewContent;
+import open.furaffinity.client.fragmentTabsOld.notImplementedYet;
+import open.furaffinity.client.fragmentTabsOld.webViewContent;
 import open.furaffinity.client.pages.msgPmsMessage;
 import open.furaffinity.client.utilities.messageIds;
