+++ app/src/main/java/open/furaffinity/client/adapter/viewSectionsPagerAdapter.java
@@ -12,12 +12,12 @@ import androidx.fragment.app.FragmentPagerAdapter;
 import java.util.ArrayList;
 
 import open.furaffinity.client.R;
-import open.furaffinity.client.fragmentTabs.comments;
-import open.furaffinity.client.fragmentTabs.notImplementedYet;
-import open.furaffinity.client.fragmentTabs.viewFolders;
-import open.furaffinity.client.fragmentTabs.viewInfo;
-import open.furaffinity.client.fragmentTabs.viewKeywords;
-import open.furaffinity.client.fragmentTabs.webViewContent;
+import open.furaffinity.client.fragmentTabsNew.comments;
+import open.furaffinity.client.fragmentTabsOld.notImplementedYet;
+import open.furaffinity.client.fragmentTabsOld.viewFolders;
+import open.furaffinity.client.fragmentTabsOld.viewInfo;
+import open.furaffinity.client.fragmentTabsOld.viewKeywords;
+import open.furaffinity.client.fragmentTabsOld.webViewContent;
 import open.furaffinity.client.pages.view;
 import open.furaffinity.client.utilities.messageIds;
