+++ app/src/main/java/open/furaffinity/client/fragmentTabsOld/manageSubmissions.java
@@ -1,4 +1,4 @@
-package open.furaffinity.client.fragmentTabs;
+package open.furaffinity.client.fragmentTabsOld;
 
 import android.content.res.ColorStateList;
 import android.os.AsyncTask;
