+++ app/src/main/res/layout/fragment_view_info.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragmentTabs.viewInfo">
+    tools:context=".fragmentTabsOld.viewInfo">
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
