+++ app/src/main/res/layout/fragment_web_view.xml
@@ -4,7 +4,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragmentTabs.webViewContent">
+    tools:context=".fragmentTabsOld.webViewContent">
 
     <LinearLayout
         android:layout_width="match_parent"
