+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -60,7 +60,7 @@
 
     <fragment
         android:id="@+id/nav_msg_pms_message"
-        android:name="open.furaffinity.client.fragmentTabs.msgPmsMessage"
+        android:name="open.furaffinity.client.fragmentTabsOld.msgPmsMessage"
         android:label="@string/menu_msg_pms_message"
         tools:layout="@layout/fragment_msgpmsmessage" />
